Switched to GDBusServer
Looks like gdbusserver is the right thing to implement, because
it accepts all the possible types of connection. For input, use
GIOChannel and g_io_channel_read_line, and add them to GMainLoop
with g_io_add_watch. Loop discovers the input and prints out the
result in callback function handle_input.